Registration Endpoint Logic moved to controller and middleware
